requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==6.0.1
flask-migrate==4.1.0
python-dotenv==1.2.1
pytest==9.0.1
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.7.2
tensorflow==2.20.0
imbalanced-learn==0.14.0
joblib==1.5.2
APScheduler==3.11.1
psycopg2-binary==2.9.11